Moved the drawing of syntax outside of the Document class.
