+++ src/main.cpp
@@ -38,6 +38,178 @@ bool writeFile(std::string path, std::string text)
     }
 }
 
+#define SYMBOL_ESCAPE '\\'
+#define SYMBOL_COMMENT '`'
+#define SYMBOL_NUMBER '#'
+#define SYMBOL_STRING '"'
+#define SYMBOL_LIST_START '['
+#define SYMBOL_LIST_END ']'
+#define SYMBOL_MAP_START '{'
+#define SYMBOL_MAP_END '}'
+#define SYMBOL_CALL_START '('
+#define SYMBOL_CALL_END ')'
+#define SYMBOL_NULL '?'
+
+typedef struct
+{
+    SDL_Color comments;
+    SDL_Color numbers;
+    SDL_Color strings;
+    SDL_Color lists;
+    SDL_Color maps;
+    SDL_Color calls;
+    SDL_Color nulls;
+    SDL_Color cursor;
+    SDL_Color undefined;
+} color_theme_t;
+
+void draw(std::vector<std::string> lines, SDL_Point to, SDL_Point cursor, SDL_Point bounds, color_theme_t *theme, FontSheet *font, SDL_Renderer *renderer)
+{
+    SDL_Point position = {0, 0};
+    SDL_Point min = {(cursor.x / bounds.x) * bounds.x, (cursor.y / bounds.y) * bounds.y};
+    SDL_Point max = {min.x + bounds.x, min.y + bounds.y};
+    auto escaping = false;
+    enum stream_state_t {
+        NOTHING_STREAM_STATE,
+        COMMENT_STREAM_STATE,
+        NUMBER_STREAM_STATE,
+        STRING_STREAM_STATE
+    };
+    auto state = NOTHING_STREAM_STATE;
+
+    for (auto line : lines)
+    {
+        for (auto symbol : line)
+        {
+            if (escaping)
+            {
+                escaping = false;
+            }
+            else
+            {
+                switch (symbol)
+                {
+                    case SYMBOL_ESCAPE:
+                        escaping = true;
+                        break;
+                    case SYMBOL_COMMENT:
+                        if (state == NOTHING_STREAM_STATE)
+                        {
+                            state = COMMENT_STREAM_STATE;
+                        }
+                        else if (state == COMMENT_STREAM_STATE)
+                        {
+                            state = NOTHING_STREAM_STATE;
+                        }
+
+                        break;
+                    case SYMBOL_NUMBER:
+                        if (state == NOTHING_STREAM_STATE)
+                        {
+                            state = NUMBER_STREAM_STATE;
+                        }
+                        else if (state == NUMBER_STREAM_STATE)
+                        {
+                            state = NOTHING_STREAM_STATE;
+                        }
+
+                        break;
+                    case SYMBOL_STRING:
+                        if (state == NOTHING_STREAM_STATE)
+                        {
+                            state = STRING_STREAM_STATE;
+                        }
+                        else if (state == STRING_STREAM_STATE)
+                        {
+                            state = NOTHING_STREAM_STATE;
+                        }
+
+                        break;
+                }
+
+                switch (symbol)
+                {
+                    case SYMBOL_ESCAPE:
+                        break;
+                    case SYMBOL_COMMENT:
+                        if (state == COMMENT_STREAM_STATE)
+                        {
+                            font->setColor(&theme->comments);
+                        }
+
+                        break;
+                    case SYMBOL_NUMBER:
+                        if (state == NUMBER_STREAM_STATE)
+                        {
+                            font->setColor(&theme->numbers);
+                        }
+
+                        break;
+                    case SYMBOL_STRING:
+                        if (state == STRING_STREAM_STATE)
+                        {
+                            font->setColor(&theme->strings);
+                        }
+
+                        break;
+                    case SYMBOL_LIST_START:
+                    case SYMBOL_LIST_END:
+                        if (state == NOTHING_STREAM_STATE)
+                        {
+                            font->setColor(&theme->lists);
+                        }
+
+                        break;
+                    case SYMBOL_MAP_START:
+                    case SYMBOL_MAP_END:
+                        if (state == NOTHING_STREAM_STATE)
+                        {
+                            font->setColor(&theme->maps);
+                        }
+
+                        break;
+                    case SYMBOL_CALL_START:
+                    case SYMBOL_CALL_END:
+                        if (state == NOTHING_STREAM_STATE)
+                        {
+                            font->setColor(&theme->calls);
+                        }
+
+                        break;
+                    case SYMBOL_NULL:
+                        if (state == NOTHING_STREAM_STATE)
+                        {
+                            font->setColor(&theme->nulls);
+                        }
+
+                        break;
+                    default:
+                        if (state == NOTHING_STREAM_STATE)
+                        {
+                            font->setColor(&theme->undefined);
+                        }
+
+                        break;
+                }
+            }
+
+            if (position.x >= min.x && position.x < max.x && position.y >= min.y && position.y < max.y)
+            {
+                auto text = std::string(1, symbol);
+                font->draw({to.x + ((position.x - min.x) * font->getFrameWidth()), to.y + ((position.y - min.y) * font->getFrameHeight())}, text, renderer);
+            }
+
+            position.x += 1;
+        }
+
+        position.x = 0;
+        position.y += 1;
+    }
+
+    font->setColor(&theme->cursor);
+    font->draw({to.x + ((cursor.x - min.x) * font->getFrameWidth()), to.y + ((cursor.y - min.y) * font->getFrameHeight())}, "_", renderer);
+}
+
 void drawTexture(SDL_Texture *texture, int x, int y, int width, int height, SDL_Renderer *renderer)
 {
     SDL_Rect source = {0, 0, width, height};
@@ -161,7 +333,8 @@ int main(int argc, char **argv)
 
             // Draw text editor
             to = {0, 24};
-            document.draw(to, bounds, &theme, &font, renderer);
+            SDL_Point cursor = {document.getCursorX(), document.getCursorY()};
+            draw(document.getLines(), to, cursor, bounds, &theme, &font, renderer);
 
             // Draw status bar
             for (auto width = 0; width < 640; width += 16)
@@ -171,14 +344,14 @@ int main(int argc, char **argv)
 
             to = {0, 316};
             auto line = " L "
-                + std::to_string(document.getCursorLineIndex())
+                + std::to_string(document.getCursorY() + 1)
                 + "/"
-                + std::to_string(document.getLineCount())
+                + std::to_string(document.getLinesCapacity())
                 + " | "
                 + "C "
-                + std::to_string(document.getCursorColumnIndex())
+                + std::to_string(document.getCursorX() + 1)
                 + "/"
-                + std::to_string(document.getColumnCount())
+                + std::to_string(document.getColumnsCapacity())
                 + " |";
             SDL_Color color = {125, 125, 125, 255};
             font.setColor(&color);
